Pin funcx-endpoint to an exact version of funcx
These packages are in 0.x and breaking changes may occur in any given
release. Right now, with the way the dependency is declared, if v0.3.6
of funcx makes changes which require an update in funcx-endpoint, then
existing versions of funcx-endpoint will begin failing if installed in
new environments.

Until/unless we set a strong policy around backwards compatibility of
releases, this will remain problematic and unsafe. To avoid potential
issues, `funcx-endpoint` will pin the exact version of `funcx` which
it requires. This ensures that we are able to install and run
`funcx-endpoint` v0.3.6 even if (for example) `funcx` v0.3.99 is not
compatible with it.